Fix #2068: Remove master public key from application detail
