+++ powerauth-admin/src/main/java/com/wultra/security/app/admin/controller/ApplicationController.java
@@ -16,6 +16,7 @@
 
 package com.wultra.security.app.admin.controller;
 
+import com.wultra.security.powerauth.client.model.response.v3.GetApplicationDetailResponse;
 import com.wultra.security.powerauth.client.v3.PowerAuthClient;
 import com.wultra.security.powerauth.client.model.entity.*;
 import com.wultra.security.powerauth.client.model.enumeration.CallbackUrlType;
