+++ powerauth-client-model/src/main/java/com/wultra/security/powerauth/client/model/response/v3/GetApplicationDetailResponse.java
@@ -1,0 +1,41 @@
+/*
+ * PowerAuth Server and related software components
+ * Copyright (C) 2025 Wultra s.r.o.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as published
+ * by the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>.
+ *
+ */
+
+package com.wultra.security.powerauth.client.model.response.v3;
+
+import com.wultra.security.powerauth.client.model.entity.ApplicationVersion;
+import lombok.Data;
+
+import java.util.ArrayList;
+import java.util.List;
+
+/**
+ * Model class representing response with application detail.
+ *
+ * @author Petr Dvorak, petr@wultra.com
+ */
+@Data
+public class GetApplicationDetailResponse {
+
+    private String applicationId;
+    private String masterPublicKey;
+    private List<String> applicationRoles = new ArrayList<>();
+    private List<ApplicationVersion> versions = new ArrayList<>();
+
+}
