+++ powerauth-client-model/src/main/java/com/wultra/security/powerauth/client/v4/PowerAuthClient.java
@@ -29,6 +29,7 @@ import com.wultra.security.powerauth.client.model.error.PowerAuthClientException
 import com.wultra.security.powerauth.client.model.request.*;
 import com.wultra.security.powerauth.client.model.request.v4.*;
 import com.wultra.security.powerauth.client.model.response.*;
+import com.wultra.security.powerauth.client.model.response.v4.GetApplicationDetailResponse;
 import com.wultra.security.powerauth.client.model.response.v4.*;
 import org.springframework.util.MultiValueMap;
