+++ powerauth-java-server/src/main/java/com/wultra/security/powerauth/app/server/controller/api/v3/ApplicationController.java
@@ -18,11 +18,12 @@
 
 package com.wultra.security.powerauth.app.server.controller.api.v3;
 
+import com.wultra.security.powerauth.app.server.service.behavior.tasks.v3.ApplicationDetailServiceBehavior;
 import com.wultra.security.powerauth.client.model.request.CreateApplicationRequest;
 import com.wultra.security.powerauth.client.model.request.GetApplicationDetailRequest;
 import com.wultra.security.powerauth.client.model.request.LookupApplicationByAppKeyRequest;
 import com.wultra.security.powerauth.client.model.response.CreateApplicationResponse;
-import com.wultra.security.powerauth.client.model.response.GetApplicationDetailResponse;
+import com.wultra.security.powerauth.client.model.response.v3.GetApplicationDetailResponse;
 import com.wultra.security.powerauth.client.model.response.GetApplicationListResponse;
 import com.wultra.security.powerauth.client.model.response.LookupApplicationByAppKeyResponse;
 import com.wultra.core.rest.model.base.request.ObjectRequest;
@@ -52,6 +53,7 @@ import org.springframework.web.bind.annotation.RestController;
 public class ApplicationController {
 
     private final ApplicationServiceBehavior applicationService;
+    private final ApplicationDetailServiceBehavior applicationDetailService;
 
     /**
      * Get the list of applications.
@@ -99,7 +101,7 @@ public class ApplicationController {
         final GetApplicationDetailRequest req = request.getRequestObject();
         logger.info("action: getApplicationDetail, state: initiated, applicationId: {}", req.getApplicationId());
         logger.debug("action: getApplicationDetail, state: initiated, request: {}", request);
-        final ObjectResponse<GetApplicationDetailResponse> response = new ObjectResponse<>(applicationService.getApplicationDetail(req));
+        final ObjectResponse<GetApplicationDetailResponse> response = new ObjectResponse<>(applicationDetailService.getApplicationDetail(req));
         logger.info("action: getApplicationDetail, state: succeeded");
         logger.debug("action: getApplicationDetail, state: succeeded, response: {}", response);
         return response;
